+++ src/app/locales/en/components/modals.js
@@ -1,0 +1,5 @@
+import saveStorageModificationAckModal from './modals/save-storage-modification-ack-modal';
+
+export default {
+  saveStorageModificationAckModal,
+};