+++ src/app/locales/en/utils/storage-actions/save-storage-modification-action.js
@@ -1,0 +1,5 @@
+export default {
+  successNotificationText: 'Storage backend "{{storageBeforeModification.name}}" has been modified successfully.',
+  failureNotificationActionName: 'modifying storage backend',
+  warningStorageCheckFailed: 'File read/write test has failed after "{{name}}" storage backend modification. Please make sure that the storage backend configuration is correct.',
+};